requirements.txt
streamlit
groq
python-dotenv
moviepy
langdetect